feat: 멀티배팅 시스템 완성 및 익스체인지 UI 개선
- 멀티배팅 API 완성 (ExchangeOrders 테이블 활용)
- 익스체인지 사이드바 금액 입력란을 스포츠북과 동일하게 수정
- 데이터베이스 스키마 수정 (potentialProfit을 DECIMAL로 변경)
- 버튼 클릭으로 멀티배팅 선택 (스포츠북과 동일한 UX)
- 노란색 버튼 하이라이트로 선택 상태 표시
- 인증 토큰 처리 개선

+++ server/models/exchangeOrderModel.js
@@ -27,8 +27,8 @@ const ExchangeOrder = sequelize.define('ExchangeOrder', {
   matchedOrderId: { type: DataTypes.INTEGER, allowNull: true },
   // 거래 내역 추적
   stakeAmount: { type: DataTypes.INTEGER, allowNull: false }, // 베팅 금액
-  potentialProfit: { type: DataTypes.INTEGER, allowNull: false }, // 잠재적 수익
-  actualProfit: { type: DataTypes.INTEGER, allowNull: true }, // 실제 수익 (정산 후)
+  potentialProfit: { type: DataTypes.DECIMAL(10, 2), allowNull: false }, // 잠재적 수익 (소수점 2자리)
+  actualProfit: { type: DataTypes.DECIMAL(10, 2), allowNull: true }, // 실제 수익 (정산 후, 소수점 2자리)
   settledAt: { type: DataTypes.DATE, allowNull: true }, // 정산 시간
   // 게임 연동 필드들
   homeTeam: { type: DataTypes.STRING, allowNull: true },
@@ -113,7 +113,7 @@ const ExchangeOrder = sequelize.define('ExchangeOrder', {
     comment: '총 배당률 (멀티배팅용)'
   },
   potentialWinnings: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.DECIMAL(10, 2),
     allowNull: true,
     comment: '잠재적 수익 (멀티배팅용)'
   },
